microservices/llm_service/requirements.txt
fastapi==0.65.2
uvicorn==0.14.0
python-multipart==0.0.5
fireo==1.4.1
typing==3.7.4.3
kubernetes==11.0.0
# google-cloud-storage==2.0.0
google-cloud-logging==3.1.2
typing-extensions==4.0.1
firebase-admin==4.5.2
protobuf==3.19.4
google-cloud-secret-manager==2.11.1
langchain==0.0.118